Fix assert in *Map aggregate functions
