+++ tests/queries/0_stateless/01422_map_skip_null.reference
@@ -1,0 +1,7 @@
+([],[])
+([],[])
+([],[])
+([2],[11])
+([2],[22])
+([2],[33])
+([2],[33])
